fix issue with CP437 text codec on some systems
Most Qt5 builds include a CP437 QTextCodec nowadays,
so I didn't notice the included QCodePage437Codec did not work
since the refactoring in commit 19450b2 .

Now the includec codec is loaded if the system or Qt do not provide a
CP437 codec already. Thus fixing bp-nfoview not working on systems that
do not provide it (OS X *cough*).